backend/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.1
idna==2.10
kombu==4.6.11
Markdown==3.2.2
prompt-toolkit==3.0.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
